added an actionListener and actionPerformed which shows a show message dialog
also in order for our results to be displayed, we must first compute for the total pricce after the discount. What I did was traverse through the menu array list where I used a switch case statement. The discount array list is used to display the final result of the price.